+++ MELA-Core/MELA-DataService/src/main/java/at/ac/tuwien/dsg/mela/dataservice/api/ConfigurationService.java
@@ -217,15 +217,15 @@ public class ConfigurationService {
     @GET
     @Path("/{serviceID}/historicalmonitoringdata/ininterval/xml")
     @Produces("application/xml")
-    public MonitoredElementMonitoringSnapshots getAllAggregatedMonitoringDataInTimeInterval(@PathParam("serviceID") String serviceID, @QueryParam("startTimestamp") int startTimestamp,
-            @QueryParam("endTimestamp") int endTimestamp) {
+    public MonitoredElementMonitoringSnapshots getAllAggregatedMonitoringDataInTimeInterval(@PathParam("serviceID") String serviceID, @QueryParam("startTimestamp") long startTimestamp,
+            @QueryParam("endTimestamp") long endTimestamp) {
         return collectionService.getAggregatedMonitoringDataInTimeInterval(serviceID, startTimestamp, endTimestamp);
     }
 
     @GET
     @Path("/{serviceID}/historicalmonitoringdata/fromtimestamp/xml")
     @Produces("application/xml")
-    public MonitoredElementMonitoringSnapshots getAllAggregatedMonitoringDataFromTimestamp(@PathParam("serviceID") String serviceID, @QueryParam("timestamp") int timestamp)
+    public MonitoredElementMonitoringSnapshots getAllAggregatedMonitoringDataFromTimestamp(@PathParam("serviceID") String serviceID, @QueryParam("timestamp") long timestamp)
     {
         return collectionService.getAllAggregatedMonitoringDataFromTimestamp(serviceID, timestamp);
     }
